Don't query Sabre parent nodes
When dealing with file access, the old DAV had a shortcut way to
directly access a file node without having to traverse all parents.
The new DAV endpoint needs this as well:

- added \OCA\DAV\Tree implementation with shortcut for "files"
- disable DAVACL plugin as it would also query all parents

There are several reasons for avoiding to query the parents:
- it causes file cache access for each node, not removable as it needs
to check for existence
- with external storages and update detection, querying a folder node
causes update detection to kick in and the scanner to do its work, the
scanner will set additional locks that might needlessly prevent
concurrent access.